pyproject.toml
[tool.poetry]
name = "yimba-api"
version = "0.1.0"
description = ""
authors = ["haf <flavienhugs@incubtek.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
facebook-scraper = "^0.2.59"
apify-client = "^1.6.1"
python-slugify = "^8.0.1"
vadersentiment = "^3.3.2"
newsapi-python = "^0.2.7"
google-search-results = "^2.4.2"
wordcloud = "^1.9.3"
matplotlib = "^3.8.3"
sentry-sdk = {extras = ["fastapi"], version = "^2.13.0"}
pydantic-settings = "^2.4.0"
fastapi-pagination = "^0.12.26"
beanie = "^1.26.0"
fastapi = {version = "^0.111.0", extras = ["standard"]}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^6.1.0"
ipython = "^8.21.0"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.docker
    | \*.egg-info
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.poetry.scripts]
yimba = "src.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
